feat: add favorite and view favorite api
