+++ backend/src/controllers/product.js
@@ -230,59 +230,6 @@ class productApi {
             });
         }
     }
-
-    async collectProduct(req, res) {
-        try {
-            const { user_id, tid } = req.body;
-
-            const payload = {
-                user_id,
-                item_id: tid,
-                save_type: "Product"
-            };
-
-            const collection = await FavoriteModel.find({ user_id, item_id: tid, save_type: "Product" });
-            if (!collection || collection.length === 0) {
-                const fav = await FavoriteModel.create(payload);
-                return res.status(201).json({
-                    success: true,
-                    message: '成功收藏商品',
-                    favId: fav._id,
-                });
-            } else {
-                await FavoriteModel.deleteMany({ user_id, item_id: tid, save_type: "Product" });
-                return res.status(200).json({
-                    success: true,
-                    message: '成功取消收藏商品',
-                });
-            }
-        } catch (err) {
-            console.error(err);
-            return res.status(500).json({
-                success: false,
-                message: '收藏/取消收藏 商品失敗',
-            });
-        }
-    }
-
-    async viewCollection(req, res) {
-        try {
-            const { user_id } = req.params;
-            const collection = await FavoriteModel.find({ user_id, save_type: "Product" });
-
-            return res.status(200).json({
-                success: true,
-                message: '查看收藏成功',
-                collection: collection,
-            })
-        } catch (err) {
-            console.error(err);
-            return res.status(500).json({
-                success: false,
-                message: '查看收藏失敗',
-            });
-        }
-    }
 }
 
 module.exports = new productApi();
